stripe_webhook/stripe_webhook_requirements.txt
flask>=3.0.0
stripe>=14.4.0,<15
google-cloud-firestore
